Implement sorting preference in sidebar and memberships. Organize folders, as things are starting to grow. Fix dropdown alignment styling.
