+++ src/util/memberships.js
@@ -1,4 +1,5 @@
 import { fetchTemplate } from './twitter.js';
+import { preferences, SORT_ALPHA } from './preferences.js';
 
 /** Fetch the list memberships for the given user in the current users' lists.
   * This is a raw HTML template that Twitter uses on the add to list modal.
@@ -36,13 +37,25 @@ export function sortMemberships(elements) {
   if (!elements) return [];
 
   const isChecked = (el) => el.querySelector('.membership-checkbox').checked;
+  const getName = (el) => el.innerText.trim().toLowerCase();
 
   // checked lists first
   var sorted = elements.sort(function (a, b) {
     var aChecked = isChecked(a);
     var bChecked = isChecked(b);
 
-    return aChecked === bChecked ? 0 : aChecked ? -1 : 1;
+    var aName = getName(a);
+    var bName = getName(b);
+
+    // sort by checked status first, then name (if preferred)
+    if (aChecked === bChecked) {
+      if (preferences.sort === SORT_ALPHA) {
+        if (aName > bName) return 1;
+        if (aName < bName) return -1;
+      }
+      return 0;
+    }
+    return aChecked ? -1 : 1;
   });
 
   return sorted;
